backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
# psycopg2-binary>=2.9.0  # Commented out for now - will be installed separately if needed
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
celery>=5.3.0
redis>=5.0.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
django-extensions>=3.2.0
